+++ src/Tools/dotnet-monitor/CollectionRules/KnownCollectionRuleActions.cs
@@ -12,5 +12,7 @@ namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules
         public const string CollectTrace = nameof(CollectTrace);
         public const string Execute = nameof(Execute);
         public const string LoadProfiler = nameof(LoadProfiler);
+        public const string SetEnvironmentVariable = nameof(SetEnvironmentVariable);
+        public const string GetEnvironmentVariable = nameof(GetEnvironmentVariable);
     }
 }