+++ documentation/releaseNotes/releaseNotes.md
@@ -1,3 +1,4 @@
 Today we are releasing the 6.0.2 build of the `dotnet-monitor` tool. This release includes:
 
-- [Here is a new feature we added and it's work item] (#737)
+- Load a profiler via a collection rule action (#781)
+- Get or Set an environment via a collection rule action (#1176)